backend/app/api/endpoints/chat.py
from fastapi import APIRouter, HTTPException, Depends
from typing import List
from sqlalchemy.orm import Session
from app.core.db import get_db
from app.models.document import Document
from app.schemas.chat import ChatRequest, ChatResponse
from app.core.config import settings
from google import genai
from google.genai import types

router = APIRouter()

try:
    client = genai.Client(api_key=settings.GEMINI_API_KEY)
except Exception:
    client = None

@router.post("/query", response_model=ChatResponse)
async def chat_query(request: ChatRequest, db: Session = Depends(get_db)):
    if not client:
        # Fallback for testing or missing API key
        return ChatResponse(answer="This is a grounded answer from the mock.")
    
    try:
        # In a real app, you would retrieve documents matching the query (RAG)
        # Using Gemini SDK, we can pass multiple file URIs or use Gemini semantic retrieval tools.
        # For this spec, we just fetch files linked from DB, but usually we would use
        # Gemini search tools. Here we simulate passing all active files as context.
        # Note: Gemini natively supports uploading files and then referencing them.
        
        active_documents = db.query(Document).filter(Document.status == "PROCESSING").all()
        # In actual implementation, we'd wait for status FAILED/ACTIVE
        # or just pass their URIs down to gemini

        contents = [request.message]
        
        # In a simple implementation, we just call generate_content
        # Let's mock a simple retrieval + generation
        response = client.models.generate_content(
            model='gemini-2.5-flash',
            contents=contents,
            config=types.GenerateContentConfig(
                system_instruction="You are a helpful school assistant answering student inquiries. Answer only based on the provided documents.",
                temperature=0.2,
            )
        )
        return ChatResponse(answer=response.text)
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
