backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
google-genai==1.61.0
alembic==1.18.3
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.20


